feat(types): stabilize EventStore and Transport interfaces in synckit/types (Milestone 3)
- Add synckit/types/interfaces.go with EventStore, Transport, EventHandler alias, and Filter
- Alias synckit.EventStore and synckit.Transport to synckit/types to freeze API in one place
- Keep CursorTransport in synckit to avoid types<->cursor coupling
- Update IMPLEMENTATION_PLAN.md marking Milestone 3 complete

No functional changes; build and tests remain green.
